\name{definitions}
\alias{AC_GUMBEL}
\alias{AC_CLAYTON}
\alias{AC_FRANK}
\alias{AC_JOE}
\alias{AC_AMH}
\alias{ML}
\alias{FML}
\alias{RML}
\alias{HAC_GUMBEL}
\alias{HAC_CLAYTON}
\alias{HAC_FRANK}
\alias{HAC_JOE}
\alias{HAC_AMH}
\alias{theta.eps}

\title{Definitions
}
\description{
Global parameters are defined. If the user is asked for a copula type, choose between \code{HAC_GUMBEL}, ..., \code{HAC_AMH}, where AMH abbreviates Ali-Mikhail-Haq or the corresponding numbers.
}
\arguments{
  \item{HAC_GUMBEL = 1}{--}
  \item{AC_GUMBEL = 2}{--}
  \item{HAC_CLAYTON = 3}{--}
  \item{AC_CLAYTON = 4}{--}
  \item{HAC_FRANK = 5}{--}
  \item{AC_FRANK = 6}{--}
  \item{HAC_JOE = 7}{--}
  \item{AC_JOE = 8}{--}
  \item{HAC_AMH = 9}{--}
  \item{AC_AMH = 10}{--}
  \item{ML = 1, FML = 2, RML = 3}{
  there are three procedures regarding the estimation of HAC. It can be chosen between Pairwise Maximum Likelihood (ML), Full Maximum Likelihood (FML) and Recursive Maximum Likelihood (RML).
} 
}
